V8 AND V4
This update have major changes .
offer.html- removed T&C section as Majority of the conditions are the same .
sh.php- Delivery prompt is been replaced with HSN CODE.Another update is that the date is editable and greetings are also changed. Furthermore, Item summation is added to the main section followed by the unit rate.